Add Trello logo. Show source logs in source type dropdown.
